Refactor: context 위치 수정으로 지갑과 연동, 버그 수정
