+++ client/src/components/VM/Operations/OperationMenu/Input.jsx
@@ -1,21 +1,10 @@
-import React, {
-  useCallback,
-  useContext,
-  useEffect,
-  useRef,
-  useState,
-} from "react";
+import React, { useContext, useEffect, useRef, useState } from "react";
 import styled from "styled-components";
 import { InputContext } from "store/InputStore";
 import { MessageContext } from "store/MessageStore";
 import { WalletContext } from "store/WalletStore";
-import { debounce, getNeededMoney } from "utils/util";
-import {
-  moneyUnitArr,
-  RETURN_COUNT,
-  reversedMoneyUnitArr,
-} from "constants/constants";
-import { getAddedWallet } from "./Return";
+import { getNeededMoney } from "utils/util";
+import { moneyUnitArr, reversedMoneyUnitArr } from "constants/constants";
 import { TimerContext } from "store/TimerStore";
 
 export default function Input() {
@@ -26,33 +15,13 @@ export default function Input() {
   const walletContext = useContext(WalletContext);
   const { wallet, setWallet } = walletContext;
   const timerContext = useContext(TimerContext);
-  const { timer, setTimer } = timerContext;
+  const { setTimer } = timerContext;
 
   const [text, setText] = useState(input);
   const [cursor, setCursor] = useState(true);
   const [temp, setTemp] = useState(null);
   const [isClicked, setIsClicked] = useState(false);
   const inputRef = useRef();
-
-  const returnFn = useCallback(
-    debounce((currentInput) => {
-      setWallet(
-        getAddedWallet(wallet, getNeededMoney(currentInput, moneyUnitArr))
-      );
-      setMessage((prev) => [
-        ...prev,
-        `잔돈 ${currentInput}원이 반환되었습니다 \n`,
-      ]);
-      setInput(0);
-    }, RETURN_COUNT),
-    []
-  );
-
-  useEffect(() => {
-    if (!timer) return;
-    returnFn(input);
-    setTimer(false);
-  }, [timer]);
 
   function onChangeHandler(e) {
     setText(e.target.value);
@@ -85,7 +54,6 @@ export default function Input() {
       setCursor(false);
       return;
     }
-    // setTimer(true);
     const [newWallet, balanced] = validateInput(value, moneyUnitArr, wallet);
 
     if (newWallet) {
@@ -110,7 +78,7 @@ export default function Input() {
   function outsideClickHandler(e) {
     if (!inputRef.current.contains(e.target) && isClicked) {
       setIsClicked(false);
-      setInput(temp);
+      if (input !== 0) setInput(temp);
       setText(temp);
     }
   }
@@ -121,7 +89,7 @@ export default function Input() {
 
   useEffect(() => {
     setText(input);
-    if (input !== 0) setTimer(true);
+    setTimer(true);
   }, [input]);
 
   useEffect(() => {
